Avg for grupper, jtds-profil

+++ src/main/java/no/kantega/frisktblodtilhodet/web/HighscoreController.java
@@ -66,14 +66,14 @@ public class HighscoreController {
 
     @RequestMapping(value = "/Grupper/{aktivitet}")
     public String getForAktivitetForGrupper(@PathVariable Aktivitet aktivitet, Model model){
-        Map<Gruppe, Integer> scores = highscoreService.getGrupperAndScoreForAktivitet(aktivitet);
+        Map<Gruppe, Double> scores = highscoreService.getGrupperAndScoreForAktivitet(aktivitet);
         model.addAttribute("personAndCount", scores);
         return "highscore/list";
     }
 
     @RequestMapping(value = "/Grupper/Alle")
     public String getForAlleForGruppe(Model model){
-        Map<Gruppe, Integer> scores = highscoreService.getGrupperAndScoreForAlle();
+        Map<Gruppe, Double> scores = highscoreService.getGrupperAndScoreForAlle();
         model.addAttribute("personAndCount", scores);
         return "highscore/list";
     }
